[anchor] Rename PositionFallbackData to OutOfFlowData
This prepares for the interleaving anchor() evaluation
(Issue 1510829).

For interleaving anchor(), we'll update the style from out-of-flow
layout even when position fallback isn't used, and it will be
necessary to store some invalidation information on Element
like we currently do for @try. Since this data isn't necessarily
related to position-fallback, the class is renamed.

Bug: 1510829
Change-Id: Ie8711abcfd9b7bba44a26e7836357e7d3b05eb51
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5237624
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252603}
NOKEYCHECK=True
GitOrigin-RevId: b73f2a726d72e997a6687ec2b2385e6cd1feb32d