[A11y] Better aria-owns validity check when repairing missing parent
Fix the new test by doing a complete validity check of a potential
aria-owns relation before creating child with the aria-owns parent
as its parent. It's possible to do a complete check, including a
check for cycles, because the validity check only requires
a child Node (does not require a candidate AXObject). The test had
failed before because one of the places where we checked for
aria-owns validity did not check for a cycle, due to that check
needing an AXObject.

The new Node-based validity check allows for improvements in other
places that used it after guessing that an aria-owns relation was
going to be good, and reversing it after finding out it was bad.
Now, the check can be done beforehand and there's nothing to "undo" --
we no longer need to create a candidate AXObject just to check whether
the aria-owns parent makes sense as a parent.

Fixed: 1518753, 1519090
Change-Id: I43764e32c44506383b7af37ab56e48d10e0bca74
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5229839
Commit-Queue: Chris Harrelson <chrishtr@chromium.org>
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Auto-Submit: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252719}
NOKEYCHECK=True
GitOrigin-RevId: 1f5d7a9479fe2f134968fc5e378dfd89973579d1