Remove destruction of the resource provider in CheckForGpuContextLost
When `CheckForGpuContextLost` is called, if it is detected that the GPU
process has crashed, the resource provider is destroyed. This seems to
be causing use-after-free issues since `CheckForGpuContextLost` is
called from `GetOrCreateResourceProvider`, resulting in unexpected uses
of a now null resource provider.

Since `CheckForGpuContextLost` also sets a flag to indicate that the
context has been lost and should be recovered, it is not necessary to
set the resource provider to nullptr at this point. Therefore, the
destruction of the resource provider has been removed.

Fixed: 1520316
Change-Id: I7c2659e874284223622753fe2b4a39806362bfb2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5238385
Commit-Queue: Andres Ricardo Perez Rojas <andresrperez@chromium.org>
Reviewed-by: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252769}
NOKEYCHECK=True
GitOrigin-RevId: 812dd783e5e140e555b6a8b8b31cdb8bdcd8988d