IDB: fix crash in cursor when execution context is disconnected.
Previously, some IDBCursor code checked the status of the
IDBDatabase which seems like it should handle the context destruction
case. However,

a) this check was not applied consistently
b) this check is not reliable
c) this check was broken

To address (a),

This change unifies the checks that various IDBCursor methods (IDL
implementations) apply before proceeding with the implementation.

To explain and address (b),

Since news of the destruction of an ExecutionContext is propagated
non-instaneously, it's possible for the ExecutionContext to be
destroyed but for `ContextDestroyed()` not to have been called yet
on any given `ContextLifecycleObserver()`. For this reason, it's
not safe to rely on secondary state such as the status of the
database connection to the browser process to indicate whether
the context has been destroyed. This change double checks
`IsContextDestroyed()` before allowing operations like `continue()`.

To explain and address (c),

The database check, where it was applied, had been broken by
crrev.com/f193122784d956cdd8feb which combined the IDBDatabase and
WebIDBDatabase. The WebIDBDatabase might have previously been null
when disconnected, but the IDBDatabase is never null, so null checks
(in IDBCursor and beyond) are replaced with a call to a new method
`IDBDatabase::IsConnectionOpen()` which checks the mojo connection.
The mojo connection automatically resets on context destruction.

Bug: 1514569, 1519989
Change-Id: I9e670c03a3d596e39ef2ff14fcadb745a357def4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5218533
Reviewed-by: Joshua Bell <jsbell@chromium.org>
Commit-Queue: Evan Stade <estade@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252770}
NOKEYCHECK=True
GitOrigin-RevId: d1ce6690e99f82e8fb2f002a0da53ab185f184ae