Switch AutoSizeLazyLoadedImages back to experimental
This change switches the AutoSizeLazyLoadedImages flag back to
experimental because of several bug reports for M121:
https://crbug.com/1521479
https://crbug.com/1522175
https://crbug.com/1522116
https://crbug.com/1522027

The issues reported are because of the change to auto sizes img
UA style rules defined in:
https://html.spec.whatwg.org/#img-contain-size

That was checked in with:
https://chromium-review.googlesource.com/c/chromium/src/+/5050963

Since several sites seem to be using `size=auto` without defining width
and height attributes for images, the UA style rule is applied causing
the images to not render as expected. To avoid this the UA style rule is reverted as well.

There is a note mentioned in the html spec addressing this:
https://html.spec.whatwg.org/#sizes-attributes
"In addition, it is strongly encouraged to specify dimensions using the
width and height attributes or with CSS. Without specified dimensions,
the image will likely render with 300x150 dimensions because
sizes="auto" implies contain-intrinsic-size: 300px 150px in the
Rendering section."

R=szager

Bug: 1522175
Change-Id: I5b4d04b0660528999196e5387beac2c3d63df277
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5241231
Reviewed-by: Stefan Zager <szager@chromium.org>
Commit-Queue: Traian Captan <tcaptan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1252944}
NOKEYCHECK=True
GitOrigin-RevId: 815ec12150b98f4205cab12076ee95626f784210