Remove cpp bit-fields check
This patch removes the bit-fields type check for cpp.

This rule was added to avoid a MSVC issue that it can't pack
different type bitfields, and enforced all-bools or
all-unsigned[1].

But the a mixing check was removed[2], this is a left over.

[1] https://chromium.googlesource.com/chromium/src/+/934092a0b031b8778a37c8c7665bd93d6fe134db%5E%21/third_party/WebKit/Tools/Scripts/webkitpy/style/checkers/cpp.py
[2] https://chromium.googlesource.com/chromium/src/+/693f5ad9db64c9a51cf97c655e34b394f9820a48

Bug: 1522092
Change-Id: I6867c7ec67a49223aa96ecdc61f3f287848ffa37
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5239693
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253196}
NOKEYCHECK=True
GitOrigin-RevId: 4e66dc6e6fb0ca6a36c3f3e526aa5671f97de7d3