[bfcache] Capture source location of JS execution
This CL makes the renderer side capture the source location of JavaScript execution, and pass that to the browser side via the mojo interface EvictFromBackForwardCache().

While it is passed to the browser side, the source location will still NOT be reported as a not restored reason in Devtools or the NotRestoredReasons API. There will be a follow-up CL that implements that in the browser side and updates the test BackForwardCacheBrowserTest::EvictByJavaScript to ensure that.

JavaScript execution occurs when some script accidentally runs while the document should've been frozen. So there's no feature handle for it.

This is part of the design doc that discusses the entire project:
https://docs.google.com/document/d/1ou_BksXiEbFDUIdsQmjYOELr8DWuich5QQAI6kpZiV0/edit?resourcekey=0-t6tUcCYuCaspXzc8ihikwQ&tab=t.0#bookmark=id.dc03qjeh7spz

Bug: 1513120
Change-Id: Ibacf730a4c52b356f02206b4e17ae1143c2fcdf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5200965
Commit-Queue: Kurumi Muto <mkurumi@google.com>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253205}
NOKEYCHECK=True
GitOrigin-RevId: 61c1b2090b869300b4918ae78fab23abab8bc9b9