[css-anchor-position] Style contain anchor-name
This CL implements style containment for anchor-as 'scoped to a sub-
tree' as defined in the css-contain-2 spec[1]. Tests are made tentative
as this is currently a CSSWG resolution, not in any spec yet.

We now need to store all in-flow candidate anchors, not only out-of-
flow ones, because the selected anchor depends on whether the querying
element is in the same contain:style subtree or not.

[1] https://drafts.csswg.org/css-contain-2/#property-scoped
-to-a-sub-tree
[2] https://github.com/w3c/csswg-drafts/issues/9100

Bug: 1518526
Change-Id: I745245eb22ccb4af7aa5cfd814af2955077085b7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5237173
Commit-Queue: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253236}
NOKEYCHECK=True
GitOrigin-RevId: 3f9c4d331fc5f555578a4a69ad04ee98380a55e6