Rename BoxType() to GetBoxType() in FragmentBuilder.
When NGBoxType in PhysicalFragment was renamed to BoxType, the getter
had to be renamed, and was named GetBoxType().
Rename similarly in FragmentBuilder, for consistency. This should also
make it slightly less likely to mistakenly use the expression BoxType()
in the context of PhysicalFragment (which now means something completely
different).

No behavior changes intended.

Change-Id: Ia16d37c783c8c8caa303d4694d2a550ae7e6b8a1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5238439
Commit-Queue: Morten Stenshorne <mstensho@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253589}
NOKEYCHECK=True
GitOrigin-RevId: f40ff5146e16e183687d332d2ee0bab627cb8b19