bindings: Reuse inline storage for wrapper references
The main worker and worklet worlds are similar to the main world on
the main rendering thread. As such, they can use the inline storage
for ScriptWrappable's wrapper reference.

This CL is merely a refactoring and does not enable inline storage.

Drive-by: Add comments what the respective world types mean and
whether they are considered an isolated world.

Bug: chromium:1520621
Change-Id: I5900a9117cfed1f5078bdafe0571882ef308281d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5240531
Commit-Queue: Michael Lippautz <mlippautz@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253775}
NOKEYCHECK=True
GitOrigin-RevId: 4be190b76def0506474658634699421062c4338d