Add WebSocketError for WebSocketStream
WebSocketError is a subclass of DOMException that has `closeCode` and
`reason` attributes which can be set by the constructor.

A WebSocketError object can be passed to the writer abort() and reader
cancel() methods to close the WebSocket with a specific `closeCode` and
`reason`.

`closeCode` is used instead of `code` because DOMExceptions already have
a `code` member with a different meaning. Change the WebSocketCloseInfo
dictionary `code` member to `closeCode` for consistency.

Reject promises with WebSocketError.

BUG=983030

Change-Id: I4af4eb1520b98e3f13443cc1b738a64e67a6e023
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5237827
Reviewed-by: Domenic Denicola <domenic@chromium.org>
Commit-Queue: Adam Rice <ricea@chromium.org>
Reviewed-by: Nidhi Jaju <nidhijaju@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253781}
NOKEYCHECK=True
GitOrigin-RevId: 6c9f5175c9237ad9764e67afc9ffd08d8e45693a