[blink/canvas] Add a feature to reclaim transfer memory when hibernating
To transfer the snapshot of an hibernated canvas, shared memory must be
allocated. It's not freed until the next frame, which thus wastes memory
since a renderer hosting an hibernating canvas is in background, and not
producing frames.

This CL adds a feature to address this.
On a 14 inch macbook pro, this saves 29MiB for a single google docs tab,
when canvas hibernation is enabled, which is not the case by default.
Similar (albeit smaller, since this depends on screen resolution, unless
the client has a High-DPI screen) savings should be observed on other
platforms.

Bug: 1511658
Change-Id: Ic685f6d798c28e3bd84c7d9457ca39a3069930f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5130231
Reviewed-by: Aaron Krajeski <aaronhk@chromium.org>
Reviewed-by: Ian Vollick <vollick@chromium.org>
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Auto-Submit: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Jonathan Ross <jonross@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253799}
NOKEYCHECK=True
GitOrigin-RevId: e42934f09e71ba67404a7240f696255f47d2aeb6