Add running status tests for the sw static routing API
This CL adds tests for the router condition, which is runningStatus.
The runningStatus condition depends on the service worker's running
status, which is not exposed to JavaScript. The only way to update the
running status is using internal API. So this CL puts test files into
non external directory.

Bug: 1519727
Change-Id: Id49f3658f85535b24d6fc491860a823cb67b1528
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5248487
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253807}
NOKEYCHECK=True
GitOrigin-RevId: 4d644c404364efc5843d629f63b9bd6069f017c1