[anchor] Pass available PhysicalSize to AnchorEvaluatorImpl constructor
The current API relies on the "just-in-time" SetAxis call to provide
the size along the relevant axis. However, this API is hard to satisfy
when resolving anchor() functions by interleaving style updates from
layout (crbug.com/1510829), as by the time we know the relevant axis,
we're deep in the style code, and we have no idea what the available
size is.

Hence, this CL instead passes the full PhysicalSize to the
AnchorEvaluatorImpl in the first place, and then relies on is_y_axis
to select either the width or the height evaluation-time.

There should be no behavior change in this CL. It's a preparation
that makes it easier to resolve anchor() functions during style
resolution.

Bug: 1510829
Change-Id: Iab2ff2edf7b9b1d0ed44d08d51d69af3bab16799
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5237607
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Reviewed-by: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1253890}
NOKEYCHECK=True
GitOrigin-RevId: e0ba36f1bfba4459d7a3a18e5c3a649526eaedd1