vt: Snapshot containing block resize skips transition
This CL adds a check for the snapshot containing block (SCB) changing
before capturing new state. We already checked for a changed SCB on each
animation frame, however, this one corresponds to the check that
occurs in the Activate View Transition[1] steps of the spec. This
difference is web-observable since, without this line, the transition
isn't skipped until animation frames are being produced. This means we
don't reject the `ready` promise, as a strict reading of the spec would
imply.

This was also an issue in same-document view transitions (i.e. triggered
by `startViewTransition()`). This CL affects both same- and
cross-document transitions.

[1] https://www.w3.org/TR/css-view-transitions-1/#activate-view-transition:~:text=transitions%20to%20false.-,If%20transition%E2%80%99s%20initial%20snapshot%20containing%20block%20size%20is%20not%20equal%20to%20the%20snapshot%20containing%20block%20size%2C%20then%20skip%20the%20view%20transition%20for%20transition%2C%20and%20return.,-Capture%20the%20new

Bug: 1404957
Change-Id: I6fabee0f7106ba21597fcc576277beb6604dec6b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5238295
Reviewed-by: Khushal Sagar <khushalsagar@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254062}
NOKEYCHECK=True
GitOrigin-RevId: c20200b42d4fe70979fedaed10ac2061b0da0dc7