sensors: Stash virtual sensor readings sent before a sensor is activated
This is a follow-up to the discussion in https://crrev.com/c/5233627,
where it was argued that the current behavior of discarding virtual
sensor updates when the sensor is inactive or not present is not very
intuitive or user-friendly.

Updates were discarded in the following situations but not anymore:
- The virtual sensor was created, but the sensor itself not (e.g. there
  has been no call to Sensor.start()).
- The virtual sensor was created, one or more Sensor instances have been
  started, activated and they have all been stopped.
- The virtual sensor was created, one or more Sensor instances have been
  started but they have not been activated yet.
- The virtual sensor was created, one or more Sensor instances have been
  created and started, but the page has become hidden.

In all the cases above, we now stash the reading and try to submit it
once the sensor becomes active again (in //services parlance, that is
when VirtualPlatformSensor::StartSensor() is called). If multiple
updates are sent while the sensor is not active, only the last one will
be used in the delayed update.

To avoid races and out-of-order updates, all calls to
VirtualPlatformSensor::AddReading() are now asynchronous and go through
PostTask().

This change makes it easier to use the Device Orientation API with
WebDriver, as contrary to the Generic Sensors API, all sensor creation,
initialization and suspension is done implicitly via event listeners.
This meant that before this change, one needed to be very careful with
the order of virtual sensor creation, event listener creation, and
submission of reading updates. Specifically, the only reliable way to
add readings and receive them with WebDriver was to:
1. Create virtual sensors
2. Add a dummy device{motion,orientation,orientationabsolute} event
   listener
3. Use WebDriver to wait for the virtual sensors to be initialized and
   activated
4. Add one or more real event listeners
5. Send an update via WebDriver

With this CL, one only needs to:
1. Create virtual sensors
2. Send an update via WebDriver
3. Add one or more real event listeners
with steps 2 and 3 being interchangeable.

Bug: 1520919, 1506995
Change-Id: Ic560e4db82490f3641984b989a984837cd7316d2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5245506
Reviewed-by: Reilly Grant <reillyg@chromium.org>
Auto-Submit: Raphael Kubo Da Costa <raphael.kubo.da.costa@intel.com>
Commit-Queue: Reilly Grant <reillyg@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254071}
NOKEYCHECK=True
GitOrigin-RevId: f826f53fb7df5ab277b41e2b256afafa674289b9