Surface full error messages for software video encoders.
These strings are all bundled with Chromium so there's no worry
of information leak when software codecs are used.

This change also sets the `unsanitized_message` field on created
exceptions in the hopes that this would show up correctly in
the console, but it doesn't seem to work with console.{error,log}.

This updates OpenH264 encoder to ensure it rejects unsupported
subsampling and bit depth so we can drop duplicated messages from
VideoEncoder.

R=eugene

Bug: 1522959
Change-Id: I39107ca41bce2e5f84dadf585e8ae1fd168d64c3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5246631
Commit-Queue: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Eugene Zemtsov <eugene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254329}
NOKEYCHECK=True
GitOrigin-RevId: f74c31828e1c2b1ff99a9f60eae2c70570bc3ccb