[SAA] Implementing SharedWorker support (Part 1: Expose same-site cookies IPC)
The Storage Access API provides access to unpartitioned cookies in
third-party contexts. This CL is part of a series to extend that access
to SharedWorkers.

Before we actually provide access to SharedWorkers, we need a way for
the renderer to limit the scope of SameSite cookies provided to the
worker. Normally, a worker with a first-party storage key could access
all cookies, but if access is via the SAA handle we want to limit that
access to just SameSite: None cookies.

This CL is part of a series to allow that limiting ability. To start, we
just implement the new option as IPC so that can be reviewed on its own.

Note: This CL adds no new capabilities, it just adds an option which
always must be set to provide the same access that is already possible.
The ability to use kNone in a first-party context will be added in the
next CL. The ability to use kAll in a third-party context will not be
added in this series of CLs.

Explainer:
https://privacycg.github.io/saa-non-cookie-storage/shared-workers.html

Part 1: Expose same-site cookie IPC
Part 2: Expose same-site cookie IDL
Part 3: Implement same-site cookie filtering
Part 4: Add SharedWorker support to SAA handle

Bug: 1484966
Change-Id: I368d3425083a8bf39e6137968daf57948e382610
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5250470
Reviewed-by: Asami Doi <asamidoi@chromium.org>
Reviewed-by: Christian Dullweber <dullweber@chromium.org>
Reviewed-by: Joe Mason <joenotcharles@google.com>
Auto-Submit: Ari Chivukula <arichiv@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Reviewed-by: Mike Taylor <miketaylr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254455}
NOKEYCHECK=True
GitOrigin-RevId: 5521d5ed2ddf2b428e668f1fa25697f4a3d70ff4