Revert "Exclude scroll margin from ScrollIntoView visibility check."
This change caused incorrect scrolling during text editing, because it
made an assumption that the caller of ScrollRectToVisible would pass a
rect that includes the scroll margin.  That assumption is valid for
element.scrollIntoView(), but not FrameSelection::RevealSelection.

This reverts crrev.com/1228025.

Bug: 1522859,1325839
Change-Id: I80748f69c59fbe7a9002ab574be4146686bd191c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5250849
Commit-Queue: Steve Kobes <skobes@chromium.org>
Reviewed-by: David Awogbemila <awogbemila@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254479}
NOKEYCHECK=True
GitOrigin-RevId: 912da1d267541a5d4ca35c9cc6ddeef7b4c6e537