Simplify AttributionSrcLoader test to avoid uaf
The simplified test still fulfills its use.

In a follow-up further investigation will be done to improve the test
suite and better understand the root cause of the flakiness which seems
to be linked to the data host being reset in
`MaybeRegisterAttributionHeaders`.

Bug: 1521971
Change-Id: If3b199aac77adaf16d83cd0a3d3820d670cdf582
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5249148
Reviewed-by: Andrew Paseltiner <apaseltiner@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Commit-Queue: Anthony Garant <anthonygarant@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254504}
NOKEYCHECK=True
GitOrigin-RevId: 73c4a4cadaa8058146ce9cf27c08d2eefd01d7a1