[a11y] NextOnLine/PrevOnLine computation on aria-hidden nodes
Currently, there is a bug with the following tree structure:

<div contenteditable="true">
<div>first line</div>
<span>this</span><span aria-hidden="true">is⁠</span><span> broken.</span>
<div>last line</div>
</div>

We compute "this" NextOnLine to be "is" even though it is
ignored/invisible. This logically makes no sense, since the node
is invisible so it shouldn't be another node's Next/PrevOnLine.
This also caused an issue with NVDA when navigating by line.

It would not announce the whole line when navigating to the second
line in the example, instead stopping right before the ignore node.

This was because the behavior explained above and the fact that
Next/PrevOnLine are only computed for unignored nodes, so essentially
we were taking "is" as the start of a new line since it had no
PrevOnLineId attribute set on it.

To fix the bug we added a check for if the anchor of the position
is ignored, and also made it so during the Next/PrevOnLine computation
we set it to the next non-ignored node.

Change-Id: I4113b9357fc3b762b99554f8ce16e4a70498264d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5247179
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Commit-Queue: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254831}
NOKEYCHECK=True
GitOrigin-RevId: 25f10df7e0825af1189108047ee4cc80b46528e0