Add web setting for disabling scrollbar styles
This change is part 2 of a 4-part series of changes to implement logic
for disabling custom scrollbar styling based on user preference. Custom
scrollbar styling can be problematic for accessibility when colors with
low contrast are used or the scrollbar width is too small.

The changes in this CL register the web preference and add the web
setting that can be used to implement logic to disable ::webkit prefixed
scrollbar pseudo elements as well as CSS ScrollbarColor and
ScrollbarWidth properties, based on user preference.

A subsequent change will include the logic to disable the styling, gated
behind the web preference. There will not be UI added to actually allow
a user to control that preference, but that would be a future change.

AX-Relnotes: n/a
Bug: 1012117
Change-Id: Idc6343467e827cc1f32b4d2f8a85cc7aeee4cb8d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5194928
Reviewed-by: Dominic Battre <battre@chromium.org>
Reviewed-by: Benjamin Beaudry <benjamin.beaudry@microsoft.com>
Reviewed-by: Alison Maher <almaher@microsoft.com>
Commit-Queue: Evelynn Kaplan <evelynn.kaplan@microsoft.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254848}
NOKEYCHECK=True
GitOrigin-RevId: 840e1c0e7e65065183f7256c0ed6d5fb33c7ca6d