Create Android Anchor Manager
Prototype the anchors extension on Android. Note that one struct was
missing from the original `dev` drop of xr_android that is now added.

This also changes the return type of `GetAnchorFromMojom` to allow
returning multiple types of errors. This is used in case the Android
API reports that it has stopped tracking an Anchor as a hint that we
should remove the anchor from consideration for future processing.
This matches what we do in the ArCoreAnchorManager as well as what
the spec allows.

Bug: 1435548
Change-Id: I8b1268a5e99066fd614391dc555d4ba8b774bb25
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5253988
Reviewed-by: Piotr Bialecki <bialpio@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1254880}
NOKEYCHECK=True
GitOrigin-RevId: 33fe95dfb1eb563947028baf487f572d57b2a2fc