Rename CanonicalCookie::IsSecure to SecureAttribute
CanonicalCookie::IsSecure() is a simple getter that returns if a
cookie's "Secure" attribute was set or not.

I.e.: Set-Cookie: "foo=bar; Secure"

With the introduction of "source_scheme" a cookie could also be
considered secure if their setting scheme was secure. To try and help
reduce potential confusion this CL renames IsSecure() to
SecureAttribute() to hopefully indicate that this is a getter and not a
function which is trying to answer the question "was this cookie created
in a secure context?".

This CL is a simple renaming and does not change any logic.

Bug: 1170548
Change-Id: I01141da043604e793eed2277fe28b4bb8ff58bfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5245913
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Steven Bingler <bingler@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Reviewed-by: Kenichi Ishibashi <bashi@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255100}
NOKEYCHECK=True
GitOrigin-RevId: 59de2a5281526bdd3380d2449b46c92ed9ce34f3