Log to Devtools browser-side failed trigger registrations
With the in-browser migration, triggers are now parsed in the browser.

With this change:
- When a parsing error is encountered in the browser, a Devtools issue
  is logged and an event is emitted to the
  Conversions.TriggerRegistrationError9 histogram.

- When a non suitable reporting origin is encountered, a Devtools issue
  is logged.

This CL brings issue logging from the browser on par with issue logging
from blink. The `http/tests/inspector-protocol/attribution-reporting` WPT tests suite now runs with the in-browser migration enabled.

Bug: 1457238
Change-Id: Id41bd516c6f8c3b2033e13fd9ab3dc17ab24fba8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4674567
Reviewed-by: Nan Lin <linnan@chromium.org>
Commit-Queue: Anthony Garant <anthonygarant@chromium.org>
Reviewed-by: Dominic Farolino <dom@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255137}
NOKEYCHECK=True
GitOrigin-RevId: 219580efaddfd009d43847d205a32d03175b8ea8