Create documentation on rolling unrar
Some of this might be obvious to people with lots of github
experience, but I don't think that's a given for Chrome
developers. This is the process I followed to create
https://github.com/aawc/unrar/pull/38. I will update the process for
completing a roll in the chromium repo once I do that for 6.2.12.

Bug: 1372527
Change-Id: Iffadf65516e3c621ad2f7759c03aabb83883daa2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5249766
Commit-Queue: Varun Khaneja <vakh@chromium.org>
Reviewed-by: Varun Khaneja <vakh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255158}
NOKEYCHECK=True
GitOrigin-RevId: 6a4370e2dd49ff6e142292589956f25a2fc7b9bf