Prototype Unbounded Space on Android
Creates an UnboundedSpaceProvider for Android as well as a new test page
for unbounded scenarios. The page itself is forked from the phone-ar
page, with the only changes being made to add checkboxes to allow
toggling between 'immersive-vr' and 'immersive-ar' and to change the
former pages use of 'local' with 'unbounded' exclusively.

Bug: 1435548
Change-Id: I6f8b79074190892330cc3700d6223ead2940a5bd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5254121
Reviewed-by: Brandon Jones <bajones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255390}
NOKEYCHECK=True
GitOrigin-RevId: 698556f9e68b4f9f0cc1bfe6c1183aa05bf756d0