Avoid CHECKing for missing scrollable area
It seems the node/layout_object returned from a hit test is sometimes
not the one we used to hit test a scrollable area. Not sure how this
happens but we should avoid crashing in this case and return null which
will just use style from the hit layout object to style a cursor (which
is what was done prior to https://crrev.com/a7b66a751c2ed55).

I'll add crash keys to debug this in a followup.

Bug: 1519197
Change-Id: I2afe75f9ce7287492bb5e80677b6dc3cb603582d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5259180
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: David Bokan <bokan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255556}
NOKEYCHECK=True
GitOrigin-RevId: 1cde40e66f3e7e16768c259e760767a354c79e06