Add crash keys for scroll corner crash
This CL adds crash keys to record the type of LayoutObject when a hit
test hits a scroll corner but the hit test's hit node doesn't have a
ScrollableArea. Two crash keys are added to record the LayoutObject on
the ScrollableArea, at the time the scroll corner hit is recorded, and
the layout object that's being used to compute scrollbar style (which
comes from `HitTestResult.InnerPossiblyPseudoNode()`).

Bug: 1519197
Change-Id: I109f3dcd14aeb201b02a2fc39ece52bb7e156d05
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5259501
Commit-Queue: David Bokan <bokan@chromium.org>
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255557}
NOKEYCHECK=True
GitOrigin-RevId: 29d3c59a3de090dc65f1f9ac26b01b52d85b54ce