Don't retarget nullptr in InvokeEvent::invoker
This fixes a bug where dispatching an InvokeEvent with a null invoker
will cause a nullptr exception as it tries to retarget the Invoker based
on the currentTarget.

The fix returns early nullptr if invoker is a nullptr, which also skips
the CHECK_EQ on eventPhase(), as just guarding the Retarget call would
skip to this check which would cause the check to fail if invoker was
retrieved during the bubbling phase.

Bug: 1511341
Change-Id: I6e0e250a47466a502daeb8ab32335ff1521edc48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5260418
Commit-Queue: Joey Arhar <jarhar@chromium.org>
Auto-Submit: Keith Cirkel <chromium@keithcirkel.co.uk>
Reviewed-by: David Baron <dbaron@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255591}
NOKEYCHECK=True
GitOrigin-RevId: ceb211accce50716d8fbc3cf2bd14ac95ebf0936