Use base::ProtectedMemory to harden MojoJS enablement
This change uses base::ProtectedMemory, which was restored in
https://crrev.com/c/5161001 to harden MojoJS enablement against data-
only attacks.

The change adds a new static member to blink::ContextFeatureSettings
that is a base::ProtectedMemory<bool>. A method to set this value to
true is added, and is called in RenderFrameImpl in all code paths that
enable MojoJS for a RenderFrame.

Finally, the isMojoJSEnabled method of blink::ContextFeatureSettings,
is augmented to CHECK that the new base::ProtectedMemory<bool>
value is true if enable_mojo_js_ is true.

The result is that if a data-only attack is used to directly enable
MojoJS outside of the method calls, the CHECK will fail and crash the
process.

Tests to verify that all the normal code paths continue to properly
enable MojoJS, as well as tests that verify that tampering with the
vulnerable properties results in a crash have been added.

Bug: 976506
Change-Id: If4f8f856edab0e60ac7e48ad65ac759aa24b2ebc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5218835
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Jeffrey Gour <jegour@microsoft.com>
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255826}
NOKEYCHECK=True
GitOrigin-RevId: 70608c6d5cdf14645fb12eae8c4b62bdb7f2623e