Preserve opened layer content across canvas flushes
When the canvas has to be presented while layers are opened, we cannot
present the layer content, but we can present any painting that happened
before the top most layer was opened. To do so, the recording has to be
split into two parts: a main recording and a side recording. The side
recording holds the content of all nested layers in one single
recording. When the top most layer is closed, the content of the side
recording is written to the main recording.

Bug: 1484741
Change-Id: Id32412819c0395edd501243b2ccead30ebc0b674
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5113500
Commit-Queue: Jean-Philippe Gravel <jpgravel@chromium.org>
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255834}
NOKEYCHECK=True
GitOrigin-RevId: f6504f4305e47cd1871261b593eb99757e092857