Add MP4 mimetype to the existing MediaRecorder web tests.
The change adds MP4 mime type to the relevant tests
on the MediaRecorder tests where the test reached to the
muxer level.

There is a bug that block the change of 'PMediaRecorder-peerconnection.https.html', which will be handled with
bug fix of 1523375.

Bug: 1072056
Change-Id: Ie3c2c306bb454e1eb8914047137c6a5844bcdfdf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5250746
Commit-Queue: Sunggook Chue <sunggch@microsoft.com>
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1255836}
NOKEYCHECK=True
GitOrigin-RevId: 57876f66c6c16a53d12d9eade0c0639879e65473