[A11y] Do not include hidden nodes when they are ignored
The only time we need hidden nodes is when they are part of a
label or description. This CL keeps track of whether a given node
has ever been part of a label or description. If the node is hidden,
it removes it from the tree unless it's needed for a label/description.

In combination with CL:4852390, which removes hidden line breaks,
reduces the time spent in serialization and deserialization by
massive amounts in tests that involve heavy usage of hidden subtrees.
For the specific results, see:
https://docs.google.com/spreadsheets/d/1ozLJMRT32yIxE0ja0rDhdZPb76hQtDd0klcRcR_uWYU/edit?pli=1#gid=0
This will make a significant difference in web apps with panels,
chats, slides, etc. where a significant percentage of the content
is hidden.

Bug: 1448706, 1353788
Cq-Include-Trybots: luci.chromium.try:fuchsia-x64-accessibility-rel,linux-blink-web-tests-force-accessibility-rel
Change-Id: I1b0de1d9b6f037c107efd3ca93c073382fd0ece1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4574033
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Chris Harrelson <chrishtr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1256230}
NOKEYCHECK=True
GitOrigin-RevId: 3a824729b09333839af2c00aa2d4f1cfc914ca34