Allow streaming local compile hints if code cache requires hash check
For certain scenarios, such as WebUI and webviews, code cache data is
provided to Blink before it has been validated to actually match the
loaded script data. In those scenarios, Blink is responsible for
checking that the hash of the script text matches the hash stored in the
code cache entry, and not using the code cache data if it doesn't match.
When LocalCompileHints is enabled, the script streamer attempts to use
the hints from the code cache data before the full text of the script
has been loaded, so it can't yet compute a hash of the script text. This
is okay; using the wrong compile hints might compile a non-optimal set
of functions, but wouldn't cause incorrect behavior. This change avoids
CHECK failures when starting to use compile hints during streaming.

Bug: 1495723
Change-Id: If6a99d42fe256c3b1807ec85a684e6ee8710c46b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5238341
Commit-Queue: Seth Brenith <seth.brenith@microsoft.com>
Reviewed-by: Leszek Swirski <leszeks@chromium.org>
Reviewed-by: Marja Hölttä <marja@chromium.org>
Reviewed-by: Thomas Lukaszewicz <tluk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1256242}
NOKEYCHECK=True
GitOrigin-RevId: f86f6550e5a85b5a1e16955db6e228975a0c83d4