Add unit test verifying BMP Suite "good" test cases.
The CL at https://crrev.com/c/5253842 is blocked on Skia Gold
issues. Splitting this CL out so we can land the tests sooner and
focus on Skia Gold-specific fixes in a separate, focused CL.
This largely matches Patchset 4 of the above CL, with the requested
fixed applied and a LICENSE file added.

We now verify that BMPs classified as "good" in Jason Summers'
BMP Suite (https://entropymine.com/jason/bmpsuite/) are a pixel-
perfect match for their PNG reference equivalents, when decoded
via BMPImageDecoder.

One of the tests (pal8v4) was not pixel-perfect but is visually
equivalent at a glance. Based on the comments in the text, this
is likely due to slightly-incorrect gamma and chromacity values
in the test file, not a bug in the decoder.

Test images are public domain, created by BMP Suite 2.8.

Bug: 1523036
Change-Id: I228445124a8c4439656e3e2baa8299e80beae5f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5262799
Commit-Queue: John Stiles <johnstiles@google.com>
Reviewed-by: Peter Kasting <pkasting@chromium.org>
Auto-Submit: John Stiles <johnstiles@google.com>
Reviewed-by: Florin Malita <fmalita@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1256275}
NOKEYCHECK=True
GitOrigin-RevId: 567bd37f650084b34e79b7f985b9a9b5ddd4243a