Define new color for Fluent Overlay minimal mode thumb.
Follow up to CL:5191003

This CL defines a new ColorId for scrollbar the thumb in overlay
scrollbars when they are in "minimal mode" (i.e. not expanded).
Instead of hardcoding the color in the scrollbar theme, the color is
now properly defined using the color pipelines architecture.

Bug: 1518945
Change-Id: I1788b6868f3253b9c5db43020b1147326837c3d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5252104
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Sam Davis Omekara <samomekarajr@microsoft.com>
Reviewed-by: Yaroslav Shalivskyy <yshalivskyy@microsoft.com>
Commit-Queue: Gaston Rodriguez <gastonr@microsoft.com>
Reviewed-by: Nico Weber <thakis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1256451}
NOKEYCHECK=True
GitOrigin-RevId: adf22b11cb0e662d88a6d5df20af294442930c96