Fix failing CHECK with devtools window measuring soft-navs
There seem to be situations where a dom modification is reported to one
instance of a SoftNavigationHeuristics class, and then the paint for
that Node is reported to a different SoftNavigationHeuristics instance.

Because these are Supplements for LocalDOMWindow, we seem to be reporting painted nodes to a different window instance, specifically in cases where devtool is being rendered and the page is undergoing significant manipulation (i.e. mobile emulation, or lighthouse testing).

Change to disable soft-navs tracking for DevTools pages.  Also add proper support for nullptr being returned instead of a SoftNavigationHeuristics instance, since this was only half guarded previously.

Bug: 1523980, 1524236
Change-Id: I2d66357179be9f332baf1ad2efa2c127780a7493
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5262924
Reviewed-by: Ian Clelland <iclelland@chromium.org>
Commit-Queue: Michal Mocny <mmocny@chromium.org>
Reviewed-by: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1256459}
NOKEYCHECK=True
GitOrigin-RevId: eece40c6be163b6642550ed3ff3ad014ce4862c1